[OPIK-2559] [P SDK] Refactor search helpers and improve synchronization logic
- Moved search-related methods (`search_spans_with_filters`, `search_traces_with_filters`, and `search_and_wait_for_done`) from `helpers` to a new `search_helpers` module.
- Updated `OpikClient` to use the newly organized `search_helpers`.
- Introduced `httpx_client.READ_TIMEOUT_SECONDS` to replace hardcoded timeout values in search methods.
